Using ROOM DATABASE as cache.
