+++ app/src/main/java/hamm/android/project/data/RestaurantDatabase.kt
@@ -5,8 +5,10 @@ import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import hamm.android.project.model.Restaurant
+import hamm.android.project.model.RestaurantBase
+import hamm.android.project.model.RestaurantExtension
 
-@Database(entities = [Restaurant::class], version = 1, exportSchema = false)
+@Database(entities = [RestaurantBase::class, RestaurantExtension::class], version = 2, exportSchema = false)
 abstract class RestaurantDatabase: RoomDatabase() {
     abstract fun restaurantDao(): RestaurantDao
