+++ app/src/main/java/hamm/android/project/fragments/SplashFragment.kt
@@ -1,33 +1,21 @@
 package hamm.android.project.fragments
 
 import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.NavOptions
-import androidx.navigation.NavOptionsBuilder
 import androidx.navigation.fragment.findNavController
-import androidx.navigation.navOptions
 import hamm.android.project.R
-import hamm.android.project.viewmodels.RestaurantViewModel
+import hamm.android.project.viewmodels.MainActivityViewModel
 import hamm.android.project.viewmodels.RestaurantViewModelFactory
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.launch
 import hamm.android.project.utils.delayedCheckForLoading
 
 class SplashFragment : Fragment() {
 
-    private lateinit var mViewModel: RestaurantViewModel
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        mViewModel = ViewModelProvider(requireActivity(), RestaurantViewModelFactory(requireActivity().application)).get(RestaurantViewModel::class.java)
-    }
+    private lateinit var mMainActivityViewModel: MainActivityViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -39,11 +27,14 @@ class SplashFragment : Fragment() {
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        loading()
+        activity?.let {
+            mMainActivityViewModel = ViewModelProvider(it).get(MainActivityViewModel::class.java)
+            loading()
+        }
     }
 
     private fun loading() {
-        delayedCheckForLoading(mViewModel, 5000L) {
+        delayedCheckForLoading(mMainActivityViewModel, 5000L) {
             findNavController().navigate(
                 SplashFragmentDirections.restaurantListFragment(),
                 NavOptions.Builder()
