+++ app/src/main/java/hamm/android/project/model/RestaurantExtension.kt
@@ -1,0 +1,11 @@
+package hamm.android.project.model
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "restaurant_extension_table")
+data class RestaurantExtension(
+    val restaurantId: Int,
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    var isFavorite: Boolean = false
+)
